feat(account): add social accounts management card for easier sign-in and account security feat(account): implement passkeys management card for enhanced account security feat(account): create password management card for account setup feat(account): add sessions management card to view and revoke active sessions feat(account): implement two-factor authentication card for additional security measures refactor(account): reorganize section imports to use a consistent directory structure
